+++ Project/src/main/java/mx/LemonTrees/Project/Repository/TrabajadorRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TrabajadorRepository extends CrudRepository<Trabajador,Long>{
+public interface TrabajadorRepository extends CrudRepository<Trabajador,Integer>{
     
 }
